prefinal app :D

+++ app/src/main/java/com/ampingat/ampingatapplication/services/DownloadIntentService.java
@@ -28,8 +28,25 @@ public class DownloadIntentService extends IntentService {
     private static final String rootDir = Environment.getExternalStorageDirectory() + "/AMPiNGAT";
     private static List<VideoFile> objects;
 
+    public static volatile DownloadIntentService _instance;
+
     public DownloadIntentService() {
         super("DownloadIntentService");
+    }
+
+    public static DownloadIntentService getInstance() {
+        if (_instance == null) {
+            synchronized (DownloadIntentService.class) {
+                if (_instance == null) {
+                    _instance = new DownloadIntentService();
+                }
+            }
+        }
+
+        return _instance;
+    }
+
+    public void prepare(){
         File dirFile = new File(rootDir);
         if (dirFile.mkdir()) {
             Log.e(TAG, "created folder \"" + dirFile.getAbsolutePath() + "\"");
@@ -64,39 +81,45 @@ public class DownloadIntentService extends IntentService {
         //iterate over list of VideoMetadataModel
         for (VideoFile videoFile : this.objects) {
             Log.e(TAG, videoFile.vidname);
-            try {
-                //core download process
-                URL sourceURL = new URL(videoFile.videourl); //request url
-                URLConnection urlConnection = sourceURL.openConnection();
-                urlConnection.connect();
-                videoFile.videosize = urlConnection.getContentLength();
-                if (downloadProgressCallback != null) {
-                    downloadProgressCallback.onDownloadPrepare(videoFile.videosize, videoFile.vidname);
-                }
-
-                InputStream inputStream = new BufferedInputStream(sourceURL.openStream(), 8192);
-                OutputStream outputStream = new FileOutputStream(videoFile.videopath);
-
-                byte data[] = new byte[4096];
-                int count;
-                int currentTotal = 0;
-                while ((count = inputStream.read(data)) != -1) {
-                    currentTotal += count;
-                    outputStream.write(data, 0, count);
+            File file = new File(videoFile.videopath);
+            if (!file.isFile()) {
+                try {
+                    //core download process
+                    URL sourceURL = new URL(videoFile.videourl); //request url
+                    URLConnection urlConnection = sourceURL.openConnection();
+                    urlConnection.connect();
+                    videoFile.videosize = urlConnection.getContentLength();
                     if (downloadProgressCallback != null) {
-                        downloadProgressCallback.onDownloadProgress(count, currentTotal);
+                        downloadProgressCallback.onDownloadPrepare(videoFile.videosize, videoFile.vidname);
                     }
-                }
 
-                outputStream.flush();
-                outputStream.close();
-                inputStream.close();
 
-            } catch (IOException e) {
-                if (downloadProgressCallback != null) {
-                    downloadProgressCallback.onDownloadError();
+    //                if (!file.isFile()) {
+                        InputStream inputStream = new BufferedInputStream(sourceURL.openStream(), 8192);
+                        OutputStream outputStream = new FileOutputStream(videoFile.videopath);
+
+                        byte data[] = new byte[4096];
+                        int count;
+                        int currentTotal = 0;
+                        while ((count = inputStream.read(data)) != -1) {
+                            currentTotal += count;
+                            outputStream.write(data, 0, count);
+                            if (downloadProgressCallback != null) {
+                                downloadProgressCallback.onDownloadProgress(count, currentTotal);
+                            }
+                        }
+
+                        outputStream.flush();
+                        outputStream.close();
+                        inputStream.close();
+    //                }
+
+                } catch (IOException e) {
+                    if (downloadProgressCallback != null) {
+                        downloadProgressCallback.onDownloadError();
+                    }
+                    e.printStackTrace();
                 }
-                e.printStackTrace();
             }
         }
